requirements.txt
annotated-types==0.5.0 ; python_full_version >= "3.8.1" and python_version < "4.0"
any-api==0.1.0.9 ; python_full_version >= "3.8.1" and python_version < "4.0"
grpcio-tools==1.48.2 ; python_full_version >= "3.8.1" and python_version < "4.0"
grpcio==1.59.0 ; python_full_version >= "3.8.1" and python_version < "4.0"
jinja2==3.1.2 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
markupsafe==2.1.3 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pait==1.0.0a14 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
protobuf-to-pydantic==0.2.1 ; python_full_version >= "3.8.1" and python_version < "4.0"
protobuf==3.20.3 ; python_full_version >= "3.8.1" and python_version < "4.0"
pydantic-core==2.10.1 ; python_full_version >= "3.8.1" and python_version < "4.0"
pydantic==2.4.2 ; python_full_version >= "3.8.1" and python_version < "4.0"
setuptools==68.0.0 ; python_full_version >= "3.8.1" and python_version < "4.0"
typing-extensions==4.7.1 ; python_full_version >= "3.8.1" and python_version < "4.0"
